Dynamic ride position completed. Park size has been modelled to a 9x9 square, given that the park is 82 km^2 metres ~ 81 Hectres ~ 9km x9km size
now we need
a). dynamic ingress
b) dynamic egress (hard to do with this implementation)
c) decided whether we are looking at optimal placement and if so we may be looking at reinforcement learning algo
d) define service time for each ride (may have been defined in the
e) what are the step sizes?
f) What are we looking at optimising?